requirements.txt
python-dotenv
pytz
datetime
nextcord
